moving _compilers to tools.build.flags and other minor cleanings
